requirements.txt
python-telegram-bot==20.7 
jdatetime==4.1.0 
hijri-converter==2.3.0 
